+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,10 @@
-on: [push]
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
 
 jobs:
   paper:
