+++ .github/workflows/tests.yml
@@ -1,5 +1,11 @@
 name: Tests
-on: [push]
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
 
 jobs:
   test:
