### 2023-03-22 #5228 @Kevin-Mattheus-Moerman
@kenatcampbellmusclelab, @ToepferLab, @steinjm, @jsaucerman would you be interested in reviewing this submission to [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/about)? It concerns a Python package for automated segmentation, tracking and analysis of sarcomeres.

JOSS reviews involve reviewing a short paper, in this case [this one](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05228/joss.05228/10.21105.joss.05228.pdf).
And also evaluating/reviewing [the software](https://github.com/Sarc-Graph/sarcgraph).

For more information on what it means to review for JOSS I refer to our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

And here is a link to an example of a completed review issue, showing that the review takes place on GitHub and that reviewers are guided by a checklists: https://github.com/openjournals/joss-reviews/issues/4389

You can let me know if you are interested here. Once I have recruited enough reviewers we'll initiate the review issue.
