### 2023-03-22 #5228 @jsaucerman
Sorry I'm unable to review this article. The others you're reviewing are
great in this area. The only other relevant paper I can think of is this
one:
Cell states beyond transcriptomics: Integrating structural organization and
gene expression in hiPSC-derived cardiomyocytes - PubMed (nih.gov)
<https://pubmed.ncbi.nlm.nih.gov/34043964/>
 , where I believe Theo A Knijnenburg is the main driver on image analysis

Jeff

On Wed, Mar 22, 2023 at 6:49 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @kenatcampbellmusclelab <https://github.com/kenatcampbellmusclelab>,
> @ToepferLab <https://github.com/ToepferLab>, @steinjm
> <https://github.com/steinjm>, @jsaucerman <https://github.com/jsaucerman>
> would you be interested in reviewing this submission to the Journal of
> Open Source Software (JOSS) <https://joss.theoj.org/about>? It concerns a
> Python package for automated segmentation, tracking and analysis of
> sarcomeres.
>
> JOSS reviews involve reviewing a short paper, in this case this one
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.05228/joss.05228/10.21105.joss.05228.pdf>
> .
> And also evaluating/reviewing the software
> <https://github.com/Sarc-Graph/sarcgraph>.
>
> For more information on what it means to review for JOSS I refer to our review
> guidelines
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.
>
> And here is a link to an example of a completed review issue, showing that
> the review takes place on GitHub and that reviewers are guided by a
> checklists: #4389
> <https://github.com/openjournals/joss-reviews/issues/4389>
>
> You can let me know if you are interested here. Once I have recruited
> enough reviewers we'll initiate the review issue.
>
