+++ paper.bib
@@ -13,6 +13,7 @@
   number={3},
   pages={e1007676},
   year={2020},
+  doi={10.1371/journal.pcbi.1007676},
   publisher={Public Library of Science San Francisco, CA USA}
 }
 
@@ -24,6 +25,7 @@
   number={1},
   pages={1258},
   year={2018},
+  doi = {10.1038/s41598-018-19539-z},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -59,6 +61,7 @@ eprint = {https://www.ahajournals.org/doi/pdf/10.1161/CIRCRESAHA.116.310363}
   number={2},
   pages={C277--C283},
   year={2016},
+  doi={10.1152/ajpcell.00094.2016},
   publisher={American Physiological Society Bethesda, MD}
 }
 
@@ -70,7 +73,7 @@ number = {2},
 pages = {514-530},
 year = {2006},
 issn = {0006-3495},
-doi = {https://doi.org/10.1529/biophysj.105.070334},
+doi = {10.1152/ajpcell.00094.2016},
 url = {https://www.sciencedirect.com/science/article/pii/S0006349506722317},
 author = {Ivo A. Telley and Jachen Denoth and Edgar Stüssi and Gabriele Pfitzer and Robert Stehle},
 abstract = {To study the dynamics of individual half-sarcomeres in striated muscle contraction, myofibrils prepared from rabbit psoas muscle and left ventricles of guinea pig were immunostained with two conjugated antibody complexes consisting of a primary antibody against either α-actinin or myomesin and a secondary fluorescently labeled Fab-fragment. We simultaneously measured force kinetics and determined the positions of the Z-line and M-band signals by fluorescence video microscopy and sophisticated computer vision (tracking) algorithms. Upon calcium activation, sarcomeres and half-sarcomeres shortened nonuniformly. Shortening occurred first rapidly and exponentially during the force rise and then slowly during the force plateau. In psoas myofibrils, time-resolved displacements of the A-band in sarcomeres were observed, i.e., the two halves of individual sarcomeres behaved nonuniformly. Nonuniformity in length changes between the two halves of sarcomeres was comparable to that between two adjacent half-sarcomeres of neighboring sarcomeres. Sequential lengthening of half-sarcomeres was observed in cardiac myofibrils during the rapid phase of force relaxation. The independent dynamics of the halves in a sarcomere reveals the half-sarcomere as the functional unit rather than the structural unit, the sarcomere. The technique will facilitate the study of filament sliding within individual half-sarcomeres and the mechanics of intersegmental chemomechanical coupling in multisegmental striated muscles.}
@@ -84,6 +87,7 @@ abstract = {To study the dynamics of individual half-sarcomeres in striated musc
   number={8},
   pages={1172--1183},
   year={2019},
+  doi={10.1161/CIRCRESAHA.118.314505},
   publisher={Am Heart Assoc}
 }
 
@@ -95,6 +99,7 @@ abstract = {To study the dynamics of individual half-sarcomeres in striated musc
   number={10},
   pages={e1009443},
   year={2021},
+  doi={10.1371/journal.pcbi.1009443},
   publisher={Public Library of Science San Francisco, CA USA}
 }
 
