+++ paper.bib
@@ -37,6 +37,7 @@
   number={3},
   pages={340--347},
   year={2015},
+  doi={10.1016/j.stemcr.2015.01.020},
   publisher={Elsevier}
 }
 
