### 2023-03-22 #5228 @jsaucerman
Please remove me from this thread.

On Wed, Mar 22, 2023, 7:03 PM Saeed Mohammadzadeh ***@***.***>
wrote:

> @saeedmhz <https://github.com/saeedmhz> I am the handling editor for this
> submission and am looking for reviewers at the moment. In the mean time,
> can you work on the below?
>
>    - Can you check those potentially missing DOIs ☝️ ? You can call @editorialbot
>    check references to check them again.
>    - Can you extend your affiliation with the department and state? And
>    can you spell the USA out as the United States of America? You can call @editorialbot
>    generate pdf here to update the paper.
>
> @kenatcampbellmusclelab <https://github.com/kenatcampbellmusclelab> I
> added missing DOIs and fixed affiliations.
>
