### 2023-03-22 #5228 @saeedmhz
> Please remove me from this thread.

Hi @jsaucerman,

I apologize for any inconvenience this may have caused. Unfortunately, I am not able to remove you from the thread directly. Nevertheless, you can easily unsubscribe from the issue by visiting the issue page: https://github.com/openjournals/joss-reviews/issues/5228

On the right side of the issue page, find the "Notifications" section and click the "Unsubscribe" button.
