### 2023-03-23 #5228 @steinjm
Hi,

I'm afraid I can't help either, but I'm honored that I made the selection. In [this](https://doi.org/10.1002/cpz1.462) paper some drawbacks of the SarcGraph were briefly mentioned, mainly that relying on segmentation is either robust or user-friendly, never both.

Kind regards,
Jeroen
