### 2023-03-23 #5228 @jsaucerman
Unfortunately that requires me to log in, which is highly unusual for a
journal.

On Thu, Mar 23, 2023 at 6:17 AM Kevin Mattheus Moerman <
***@***.***> wrote:

> @jsaucerman <https://github.com/jsaucerman> if you'd like to avoid
> notifications, you can hit the unsubscribe button in the right hand side of
> GitHub to stop receiving notifications.
>
