### 2023-03-24 #5228 @Kevin-Mattheus-Moerman
@jsaucerman I see what you mean. My apologies for the inconvenience.

I think this is different from what some other journals do, which is that they may create a new account on their system for you (without your permission) and request you to sign in to accept/decline a review invite. Here the potential reviewers we invite are already active on GitHub (or at least they should ideally be), and are invited through this existing account. So the assumption is that it is easy for such potential reviewers to mute the thread, since they are already active GitHub users.
I will tag our support team here though, @openjournals/dev @arfon, to see if editors can do this for folks like yourself who prefer not to sign in.
