### 2023-04-24 Sarc-Graph/sarcgraph#2 @mbarzegary
[JOSS Review] Contribution 
Dear authors,

External contribution to the software is encouraged (in the paper), but no relevant guideline/policy is provided for it. You may consider adding a `CONTRIBUTING.md` file for this.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
