### 2023-04-24 Sarc-Graph/sarcgraph#3 @mbarzegary
[JOSS Review] Installation
Dear authors,

The software installation routine is straightforward, and there is no special issue with it. But, the general procedure was a bit weird to me. The user needs to clone the repository just for the YAML configuration file and the tutorials, and although the repository contains the source code of the software, it gets installed from PyPI online repositories. This way eliminates the possibility of installing a developer version of the tool (like for performing modification and customization to the core functionalities, the possibility of which is highlighted in the paper). Moreover, it makes the test procedure redundant because the pytest routines check the codes in the `sarcgraph` directory, but the software gets installed from another source (an online repository). You may consider adding a standard installation routine such that users can install the tool from the cloned local repository by executing `pip install -e .`.  It seems that the file `setup.py` is already there, but I’m not sure if it is functional for a local installation.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
