### 2023-04-24 Sarc-Graph/sarcgraph#4 @mbarzegary
 [JOSS Review] Reproducibility and functionality
Dear authors,

Generally speaking, the software runs as described in the documentation, and the demos/tutorials can be set up by following the instruction. However, there are minor fixes that can improve the quality of the submission in this regard:
* It can be nice to number the tutorial files so that the users can find out the order in which they need to follow the tutorials. In the online documentation, it’s not a problem because the order is known from the ToC.
* More elaboration is needed in the GPR example (time series example) to demonstrate the effect of applying GPR on the detection process. In the current state of the provided example, it's difficult to see the effect by just running the code.
* It's worth mentioning that the last tutorial (visualization) needs the output of previous tutorials to run. In the notebook file, it is only mentioned that the user should provide a set of files without saying anything about how to get them.
* Gif animations will not be displayed in the visualization notebook file. Is there any prerequisites needed for it? I checked this on both Windows and Linux, resulting in the same empty placeholder.
* A general question: why are the `x` and `y` data in a reversed order in all `plt.plot` calls that we have in tutorials? I expect it to be `plt.plot(x, y)`, but it is always the other way around. Is it not necessary to elaborate on the reason?

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
