### 2023-04-24 Sarc-Graph/sarcgraph#5 @mbarzegary
[JOSS Review] Tests
Dear authors,

I see that the software has tests implemented using pytest procedures, and they are quite comprehensive in my opinion. However, since the software installation procedure installs the software from pip repositories (and not the cloned local repo), I think it can be nice if the test routines check proper installation too. At this moment, the tests can all pass without the `sarcgraph` package being installed in the virtual environment due to the presence of the `sarcgraph` directory.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
