### 2023-04-24 Sarc-Graph/sarcgraph#6 @mbarzegary
[JOSS Review] Statement of need and the paper
Dear authors,

The paper is written well, but these issues should be addressed:
* More elaboration is needed on "The updated version is better designed, more efficient, and significantly more user-friendly" and "significant enhancements over its predecessor, such as new algorithms for improved sarcomere detection and tracking, faster computation times, and a more user-friendly library design". These sentences seem to be qualitative. Can you provide more details leading to such a conclusion?
* Why is the new algorithm more efficient than the previous one (Zhao et al. 2021)? A quantitative comparison is needed here. Comparing and elaborating on the differences between Fig 4 in this paper and Fig 6 of the previous publication can be helpful here.
* "SarcGraph is readily customizable and adaptable by the user", how is this going to come into practice? How is this customizability foreseen considering that the software comprises only 2 big source files?
* The scale bar is missing in the paper figures.
* It can be nice to mention how the reader should proceed to reproduce the figures of the paper?
* The JOSS guidelines require the paper to contain a section titled "statement of need", which is missing in this case. Part of the summary is already discussing it, but a separate section with more elaboration on the target group of users is desired. The software should also be compared to other available tools in this section.
* I think it is beneficial if you mention the limitations of the library briefly (can be in the documentation instead).

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
