### 2023-04-24 Sarc-Graph/sarcgraph#7 @mbarzegary
[JOSS Review] Documentation 
Dear authors,

The documentation is quite acceptable for JOSS in my opinion, and it is written well. However, you may consider improving it with these suggestions:
* The provided documentation lacks a theory guide. The previous publication and the original software seem to have a more comprehensive guide in this regard. You may consider adding a small appendix to the paper and/or maybe moving part of the previous publication to the documentation as a theory guide.
* The installation guide in the repo readme file is different from the hosted documentation. The readme file installs the prerequisites via a YAML file while the documentation guides the user to install them using `pip` one by one.
* You may consider removing the shell sign (`$`) from the provided commands in the installation guide (in the documentation) so that users can easily copy/run the command using the copy handle of the markdown text without needing to remove the dollar sign from the command before executing it.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/5322
