+++ docs/source/tutorial/index.rst
@@ -7,7 +7,7 @@ Demos and tutorials
 .. toctree::
    :maxdepth: 1
    
-   basics
-   timeseries
-   analysis
-   visualization
+   t1-basics
+   t2-timeseries
+   t3-analysis
+   t4-visualization
