+++ .github/workflows/tests.yml
@@ -3,9 +3,13 @@ on:
   push:
     branches:
       - main
+    paths:
+      - '**.py'
   pull_request:
     branches:
       - main
+    path:
+      - '**.py'
 
 jobs:
   test:
