+++ .github/workflows/draft-pdf.yml
@@ -2,9 +2,13 @@ on:
   push:
     branches:
       - main
+    path:
+      - '**.md'
   pull_request:
     branches:
       - main
+    path:
+      - '**.md'
 
 jobs:
   paper:
