### 2023-05-02 Sarc-Graph/sarcgraph#8 @kenatcampbellmusclelab
JOSS Review - Needs statement of need
Paper needs a clear statement of need

Reference to https://github.com/openjournals/joss-reviews/issues/5322#issuecomment-1530387975
