### 2023-05-02 Sarc-Graph/sarcgraph#9 @kenatcampbellmusclelab
JOSS Review - Functionality documentation
Although perhaps trivial for people familiar with JOSS, this software is intended to be used by biologists who are less experienced with programming.

I followed the installation instructions but had to do some googling to work out that I needed to launch a Jupyter notebook to view the tutorials. Jupyter notebooks are not part of my normal workflow and I was not familiar with the extension. I suspect many biologists would struggle to make that step.

I think a few extra lines in the readme, and perhaps even a video showing the software being demonstrated would make the software much more user-friendly.

Relates to ttps://github.com/openjournals/joss-reviews/issues/5322#issuecomment-1530387975
