### 2023-05-03 7a25f8c
docs: Separate stable/dev install, add CONTRIBUTING.md, fix pip dep install, add conda support
- Address reviewer comments to improve documentation:
  * Update installation instructions to distinguish between stable and developer versions
  * Add a new CONTRIBUTING.md file to guide contributors
- Fix the issue with pip not installing dependencies alongside sarcgraph
- Introduce conda support for easier installation of the stable version
