+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name="sarcgraph",
-    version="0.1.0",
+    version="0.1.1",
     author="Saeed Mohammadzadeh",
     author_email="saeedmhz@bu.edu",
     description="A software for sarcomere detection and tracking",
@@ -29,3 +29,16 @@ setup(
     ],
     zip_safe=False,
 )
+
+
+# install_requires=[
+#     "matplotlib==3.5.2",
+#     "networkx==2.8.4",
+#     "numpy==1.23.5",
+#     "pandas==1.5.2",
+#     "scikit-image==0.19.3",
+#     "scikit-learn==1.2.1",
+#     "scipy==1.10.0",
+#     "sk-video==1.1.10",
+#     "trackpy==0.6.1",
+# ],
