+++ docs/source/dev_guide.rst
@@ -1,0 +1,302 @@
+.. _dev_ref:
+
+**Developer's Guide**
+=====================
+
+Welcome to the Developer's Guide for SarcGraph! This guide is intended for 
+developers, contributors, and maintainers who are interested in understanding, 
+modifying, or extending the functionality of SarcGraph. Our goal is to create a 
+welcoming and collaborative environment for everyone who wants to contribute to 
+this project.
+
+In this guide, we'll cover topics such as setting up your development 
+environment, code style and conventions, testing, documentation, version control
+and branching, the contribution process, release management, and community 
+communication. We hope this information will help you become a valuable 
+contributor to the SarcGraph project.
+
+Let's dive in and get started with a project overview!
+
+**Project Overview**
+--------------------
+
+SarcGraph is a specialized software tool designed to streamline the detection, 
+tracking, and analysis of z-discs and sarcomeres in movies of beating human 
+induced pluripotent stem cell-derived cardiomyocytes (hiPSC-CMs). The main 
+objective is to offer researchers a user-friendly, adaptable, and efficient 
+solution for studying sarcomere dynamics in hiPSC-CMs.
+
+The project comprises two primary modules: sg and sg_tools.
+
+1. :ref:`sg <sg.sarcgraph>`: This module encompasses the core functionality of SarcGraph, 
+including image processing techniques for z-disc segmentation and tracking, as 
+well as sarcomere detection. The module employs the Trackpy library to handle 
+the tracking tasks effectively.
+
+2. :ref:`sg_tools <sg_tools.sarcgraphtools>`: This module features three functional categories:
+
+- :class:`TimeSeries <sarcgraph.sg_tools.SarcGraphTools.TimeSeries>`: Offers functions for handling time series data related to the 
+  sarcomeres, such as length and position information.
+
+- :class:`Analysis <sarcgraph.sg_tools.SarcGraphTools.Analysis>`: Provides post-processing analysis tools for extracting high-level 
+  mechanical properties, like average deformation or orientation order 
+  parameter.
+
+- :class:`Visualization <sarcgraph.sg_tools.SarcGraphTools.Visualization>`: Supplies a range of visualization utilities for displaying 
+  the results of detection, tracking, and analysis, enabling users to better 
+  comprehend and interpret their findings.
+
+SarcGraph's source code, written in Python, leverages various popular scientific
+computing libraries like NumPy, SciPy, scikit-image, scikit-learn, pandas, and 
+matplotlib, along with the Trackpy library for tracking purposes. The code 
+organization is simple and efficient, with each component housed in its 
+respective module.
+
+In the sections that follow, we will outline the steps for setting up your 
+development environment, working with the code, and contributing to the project.
+
+**Setting Up Your Development Environment**
+-------------------------------------------
+
+Before you can start contributing to the SarcGraph project, you'll need to set 
+up your development environment. This section will guide you through the 
+necessary steps to prepare your environment for working with the SarcGraph 
+codebase.
+
+1. **System requirements and package dependencies**
+
+SarcGraph is compatible with Linux, macOS, and Windows operating systems. We 
+have tested the software on Python 3.10, and it is recommended to use this 
+version for development. Below is a table listing the required packages and 
+their respective versions:
+
+.. list-table::
+   :header-rows: 1
+
+   * - Package Name
+       - Version
+   * - ffmpeg
+       - 4.2.2
+   * - matplotlib
+       - 3.5.2
+   * - networkx
+   * - numpy
+   * - pandas
+   * - scikit-image
+   * - scikit-learn
+   * - scipy
+   * - sk-video
+   * - trackpy
+   * - pytest
+   * - pytest-cov
+   * - flake8
+   * - black
+   * - sphinx
+   * - sphinx_rtd_theme
+   * - sphinx-design
[…]
