+++ docs/source/installation.rst
@@ -34,52 +34,71 @@ SarcGraph needs the following packages to run:
 | trackpy           | 0.6.1   |
 +-------------------+---------+
 
-**Installation Steps**
-----------------------
+**Stable Version Installation**
+-------------------------------
 
-We recommend using Anaconda to install prerequisites.
+**From Conda**
+**************
+
+We recommend using Anaconda to install sarcgraph.
 
 **1. Install Anaconda**
 
-Follow `Anaconda's installation instructions <https://docs.anaconda.com/anaconda/install/index.html>`_.
+Follow `Anaconda's installation instructions <https://docs.anaconda.com/anaconda/install/index.html>`_. Alternatively you can install `Miniconda <https://docs.conda.io/en/latest/miniconda.html>`_ (a minimal installer for conda).
 
-**2. Create a new environment**
+**2. Create a new Conda environment and install sarcgraph**
 
-Run the following command in Anaconda terminal to create a new conda environment:
+Run the following command in Anaconda terminal to create a new conda environment
+containing the sarcgraph package and all its dependencies:
 
 .. code-block:: bash
 
-    $ conda create -n sarcgraph python=3.10
+    $ conda create --name sarcgraph-env -c conda-forge -c saeedmhz sarcgraph
 
 Type ``y`` and press ``Enter`` when prompted. 
 
-Activate the ``sarcgraph`` environment:
+**3. Activate the environment**
 
 .. code-block:: bash
 
-    $ conda activate sarcgraph
+    $ conda activate sarcgraph-env
 
-**3. Install pre-requisites**
+**From PyPI**
+*************
 
-Run the following commands to install prerequisites:
+Since `ffmpeg <https://ffmpeg.org/>`_ is not availble on PyPI, you need to 
+install it seperately. We recommend using Anaconda to install the correct 
+version of ffmpeg.
 
-.. code-block:: bash
+**1. Install ffmpeg**
 
-    $ conda install -c anaconda networkx=2.8.4 numpy=1.23.5 pandas=1.5.2 scikit-image=0.19.3 scikit-learn=1.2.1 scipy=1.10.0
-
-.. code-block:: bash
-
-    $ conda install -c conda-forge ffmpeg=4.2.2 matplotlib=3.5.2 sk-video=1.1.10 trackpy=0.6.1
-
-**4. Install SarcGraph**
-
-SarcGraph can be installed using the following command:
+Install anaconda and run the following commands in Anaconda terminal to create 
+and activate a new environment:
 
 .. code-block:: bash
 
-    $ pip install sarcgraph
+    $ conda create --name sarcgraph-env python=3.10
 
-**5. Verify installation**
+.. code-block:: bash
+
+    $ conda activate sarcgraph-env
+
+Run the following command to install ffmpeg:
+
+.. code-block:: bash
+
+    $ conda install -c conda-forge ffmpeg=4.2.2
+
+**2. Install SarcGraph**
+
+SarcGraph and its dependencies can be installed using the following command:
+
+.. code-block:: bash
+
+    $ pip install --upgrade sarcgraph
+
+**Verify Installation**
+***********************
 
 Run the following command and check if it runs with no errors:
 
@@ -87,92 +106,34 @@ Run the following command and check if it runs with no errors:
 
     $ python -c "from sarcgraph.sg import SarcGraph"
 
-**Running Tests**
------------------
+**Run Tutorial Notebooks**
+--------------------------
 
-Tests in the `SarcGraph repository <https://github.com/Sarc-Graph/sarcgraph>`_ are 
[…]
