+++ doc_requirements.txt
@@ -1,15 +1,0 @@
-nbsphinx==0.8.12
-sphinx>=5.3.0
-sphinx-copybutton==0.5.1
-sphinx_design==0.3.0
-sphinx-rtd-theme==1.2.0
-
-numpy
-matplotlib
-networkx
-pandas
-trackpy
-scikit-image
-sk-video
-scikit-learn
-scipy
