+++ .readthedocs.yaml
@@ -7,9 +7,12 @@ version: 2
 
 # Set the version of Python and other tools you might need
 build:
-  os: ubuntu-22.04
+  os: "ubuntu-20.04"
   tools:
-    python: "3.10"
+    python: "mambaforge-4.10"
+
+conda:
+  environment: environment.yml
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -18,9 +21,6 @@ sphinx:
 # If using Sphinx, optionally build your docs in additional formats such as PDF
 formats:
   - htmlzip
-
-conda:
-  environment: environment.yml
 
 # Optionally declare the Python requirements required to build your docs
 # python:
