### 2023-05-04 921e6ae
Update README.md
update installation instructions
