### 2023-05-04 Sarc-Graph/sarcgraph#2 @saeedmhz
Thank you for your valuable feedback and for highlighting the need for clear guidelines for external contributions to our software.

In response to your comment, we have made the following changes to address your concerns:

1. We have introduced a "Developer's Guide" section in the documentation, which provides more in-depth information for contributors, including policies, best practices, and additional resources.

2. Furthermore, we have added a CONTRIBUTING.md file to the repository. This file provides an overview of the steps for external contributors to submit their contributions. It also directs them to the documentation for more comprehensive information on coding style, formatting, linting conventions, and testing instructions, which are detailed in the "Developer's Guide" section of the documentation.

We believe these changes will significantly improve the accessibility of our project for external contributors and foster a more collaborative environment. We appreciate your suggestion and hope that our revisions meet your expectations.

Please let us know if there are any additional concerns or suggestions.
