+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
     - name: Set up Conda
       uses: conda-incubator/setup-miniconda@v2
       with:
-        activate-environment: test-env
+        activate-environment: sarcgraph-dev
         environment-file: environment.yml
         python-version: 3.10
         auto-activate-base: false
