### 2023-05-04 Sarc-Graph/sarcgraph#3 @saeedmhz
Thank you for your insightful comments regarding the installation process of our software. We appreciate your suggestions and have made the following changes to address your concerns:

To simplify the installation process, we have uploaded our package to the conda servers, including all dependencies. This allows users to create a new environment and install everything with just one command.

We have made our tutorial files accessible on Binder, enabling users to run them without the need to set up local environments. In addition, we have provided more in-depth instructions on how to download notebooks with and without cloning the repository.

We have updated the package on PyPI servers to include all dependencies except for FFmpeg, streamlining the pip installation process for users who prefer not to use conda.

We have added a separate section for developers in the documentation, which includes instructions for an editable install using the ``pip install -e .`` command with the ``setup.py`` file in the repository. This allows developers to install a local version of the tool for modifications and customizations.

Finally, we have revised the installation instructions on the GitHub README to match the updated documentation.

We believe that these changes will improve the user experience during the installation process and make it easier for users and developers to work with our software. We hope our revisions adequately address your concerns.

Please let us know if there are any further comments or suggestions.
