### 2023-05-04 Sarc-Graph/sarcgraph#4 @saeedmhz
Thank you for your thorough examination of our software and tutorials, as well as your helpful suggestions for improving their quality. We have made the following changes in response to your comments:

1. To make the order of tutorials clearer, we have numbered the tutorial files as follows: ``t1-basics.ipynb``, ``t2-timeseries.ipynb``, ``t3-analysis.ipynb``, and ``t4-visualization.ipynb``.

2. We have expanded the explanation for the use of Gaussian Process Regression (GPR) in the timeseries tutorial (``t2-timeseries.ipynb``). We discussed the primary purposes of GPR, which are noise reduction and filling missing information, and demonstrated these benefits through additional visualizations in the tutorial.

3. In the ``t4-visualization.ipynb`` notebook, we explicitly mentioned which files should be run before executing this tutorial to ensure users have the necessary output files.

4. We have resolved the issue with GIF animations not being displayed in the visualization notebook. They should now display correctly.

5. Regarding the reversed order of ``x`` and ``y`` data in ``plt.plot calls``, this is due to the use of the ``skimage.measure.find_contours`` function, which returns contours with coordinates in ``(row, column)`` order, corresponding to ``(y, x)`` in Cartesian coordinates. As a result, we use ``y`` for the first dimension and ``x`` for the second dimension throughout the code (e.g., ``plt.plot(y, x)``). To clarify this for users, we added a warning in the ``sg.py`` file under the ``SarcGraph._detect_contours`` function.

We appreciate your attention to detail and believe that these changes will enhance the user experience and the overall quality of our software and tutorials. We hope our revisions address your concerns satisfactorily.

Please let us know if you have any further comments or suggestions.
