+++ .github/workflows/tests.yml
@@ -29,7 +29,7 @@ jobs:
       with:
         activate-environment: sarcgraph-dev
         environment-file: environment.yml
-        python-version: 3.10
+        python-version: 3.10.9
         auto-activate-base: false
 
     - name: Test with pytest
