+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SarcGraph: A Python package for automated segmentation, tracking and analysis of sarcomeres in hiPSC-CMs'
+title: 'SarcGraph: A Python package for analyzing the contractile behavior of pluripotent stem cell-derived cardiomyocytes'
 tags:
   - Python
   - Quantitative Methods
@@ -25,26 +25,40 @@ bibliography: paper.bib
 
 # Summary
 
-Heart disease remains the leading cause of death worldwide [@who_cvd_fs]. In response to this societal burden, scientific funding agencies and researchers have invested significant effort into understanding and controlling the functional behavior of heart cells. In particular, there has been a recent and growing focus on engineered heart cells and tissue to both better understand the complex interactions that drive disease, and to repair the damaged heart. An important component of these endeavors is the study of human induced pluripotent stem cell-derived cardiomyocytes (hiPSC-CMs), cells sampled non-invasively from living humans, transformed into stem cells, and subsequently differentiated into cardiomyocytes, i.e., cardiac muscle cells. These cardiomyocytes are composed of sarcomeres, sub-cellular contractile units, that can be fluorescently labeled and visualized via z-disc proteins (see \autoref{fig:intro}). One major challenge in studying hiPSC-CMs in this context is that the immaturity and structural nonlinearities of hiPSC-CMs (i.e., disordered sarcomere chain structure in comparison to the almost crystalline sarcomere structure of mature cardiomyocytes) causes significant complications for performing consistent analysis of their functional contractile characteristics. And, though multiple methods have recently been developed for analyzing images of hiPSC-CMs [@morris2020striated; @sutcliffe2018high; @pasqualini2015structural; @doi:10.1161/CIRCRESAHA.116.310363; @pasqualin2016sarcoptim; @TELLEY2006514], few are suitable for analyzing the asynchronous contractile behavior of beating cells [@toepfer2019sarctrack]. In our previous publication, we introduced a novel computational framework (`SarcGraph`) to perform this task, directly compared it to other methods, and demonstrated its state of the art efficacy and novel functionalities [@zhao2021sarc].
+Heart disease remains the leading cause of death worldwide [@who_cvd_fs]. In response to this societal burden, scientific funding agencies and researchers have invested significant effort into understanding and controlling the functional behavior of heart cells. In particular, there has been a recent and growing focus on engineered heart cells and tissue to both better understand the complex interactions that drive disease, and to repair the damaged heart. An important component of these endeavors is the study of human induced pluripotent stem cell-derived cardiomyocytes (hiPSC-CMs), cells sampled non-invasively from living humans, transformed into stem cells, and subsequently differentiated into cardiomyocytes, i.e., cardiac muscle cells. 
+These cardiomyocytes are composed of sarcomeres, sub-cellular contractile units, that can be fluorescently labeled and visualized via z-disc proteins. One major challenge in studying hiPSC-CMs in this context is that the immaturity and structural nonlinearities (i.e., disordered sarcomere chain structure in comparison to the almost crystalline sarcomere structure of mature cardiomyocytes) of hiPSC-CMs causes significant complications for performing consistent analysis of their functional contractile characteristics. And, though multiple methods have recently been developed for analyzing images of hiPSC-CMs [@morris2020striated; @sutcliffe2018high; @pasqualini2015structural; @doi:10.1161/CIRCRESAHA.116.310363; @pasqualin2016sarcoptim; @TELLEY2006514; @gerbin2021cell], few are suitable for analyzing the asynchronous contractile behavior of beating cells [@toepfer2019sarctrack]. In our previous publication, we introduced a novel computational framework (`SarcGraph`) to perform this task, directly compared it to other methods, and demonstrated its state of the art efficacy and novel functionalities [@zhao2021sarc].
+
[…]
