+++ paper.bib
@@ -118,3 +118,15 @@ abstract = {To study the dynamics of individual half-sarcomeres in striated musc
   doi          = {10.5281/zenodo.7670439},
   url          = {https://doi.org/10.5281/zenodo.7670439}
 }
+
+@article{gerbin2021cell,
+  title={Cell states beyond transcriptomics: integrating structural organization and gene expression in hiPSC-derived cardiomyocytes},
+  author={Gerbin, Kaytlyn A and Grancharova, Tanya and Donovan-Maiye, Rory M and Hendershott, Melissa C and Anderson, Helen G and Brown, Jackson M and Chen, Jianxu and Dinh, Stephanie Q and Gehring, Jamie L and Johnson, Gregory R and others},
+  journal={Cell Systems},
+  volume={12},
+  number={6},
+  pages={670--687},
+  year={2021},
+  publisher={Elsevier},
+  doi={10.1016/j.cels.2021.05.001}
+}
