+++ sarcgraph/sg_tools.py
@@ -8,6 +8,7 @@ import shutil
 import json
 import os
 
+from matplotlib.lines import Line2D
 from scipy import signal
 from scipy.signal import find_peaks
 from scipy.linalg import polar
@@ -221,14 +222,30 @@ class SarcGraphTools:
             ax.set_aspect("equal")
             ax.imshow(raw_frame[:, :, 0], cmap=plt.cm.gray)
             ax.set_title(
-                f"{len(contours)} z-disks and {len(sarcs_x)} sarcomeres\n"
-                f"found in frame {frame_num}"
+                f"{len(contours)} z-disks and {len(sarcs_x)} sarcomeres "
+                f"in frame {frame_num+1}"
             )
             for contour in contours:
                 ax.plot(contour[:, 1], contour[:, 0], "#0000cc", linewidth=1)
             ax.plot(sarcs_y, sarcs_x, "*", color="#cc0000", markersize=3)
             ax.set_xticks([])
             ax.set_yticks([])
+
+            # Create legend
+            legend_elements = [
+                Line2D([0], [0], color="#0000cc", lw=4, label="Z-disc"),
+                Line2D(
+                    [0],
+                    [0],
+                    marker="*",
+                    color="#cc0000",
+                    markersize=8,
+                    linestyle="None",
+                    label="Sarcomere",
+                ),
+            ]
+
+            ax.legend(handles=legend_elements)
 
             output_file = (
                 f"{self.sg_tools.output_dir}/zdiscs-sarcs-frame-{frame_num}"
@@ -241,6 +258,8 @@ class SarcGraphTools:
 
             if self.sg_tools.include_eps:
                 plt.savefig(f"{output_file}.eps", bbox_inches="tight")
+
+            plt.show()
 
         def contraction(self):
             """Visualize all detected sarcomeres in every frame according to
@@ -307,21 +326,28 @@ class SarcGraphTools:
             sarcs_length_norm_median = groupby_frame.length_norm.median()
             sarcs_length_norm_mean = groupby_frame.length_norm.mean()
 
+            _, ax = plt.subplots(figsize=(5, 5))
+            ax.grid("on")
             for sarc_id in range(num_sarcs):
                 indices = sarc_ids == sarc_id
-                plt.plot(sarcs_length_norm[indices], linewidth=0.25)
-            plt.plot(
-                sarcs_length_norm_median,
-                "k-",
-                linewidth=3,
-                label="median curve",
-            )
+                plt.plot(
+                    sarcs_length_norm[indices],
+                    linewidth=0.25,
+                    color=(0.545, 0.106, 0.086),
+                    alpha=0.1,
+                )
             plt.plot(
                 sarcs_length_norm_mean,
+                "k-",
+                linewidth=2,
+                label="mean curve",
+            )
+            plt.plot(
+                sarcs_length_norm_median,
                 "--",
                 color=(0.5, 0.5, 0.5),
-                linewidth=3,
-                label="mean curve",
+                linewidth=2,
+                label="median curve",
             )
             plt.xlabel("frame")
             plt.ylabel("normalized length")
@@ -343,6 +369,8 @@ class SarcGraphTools:
             if self.sg_tools.include_eps:
                 plt.savefig(f"{output_file}.eps", bbox_inches="tight")
 
+            plt.show()
+
         def OOP(self):
             """Plot recovered Orientational Order Parameter."""
             OOP = self.sg_tools._load_recovered_info("OOP")
@@ -363,19 +391,44 @@ class SarcGraphTools:
             if self.sg_tools.include_eps:
[…]
