### 2023-05-05 07b10eb
fix snippets not showing - figure 2 updated with scale bar
