+++ paper.md
@@ -92,91 +92,8 @@ In this Appendix, we present the code snippets used to generate the figures in t
 
 ![](figures/snippet1.png){width=100%}
 
-**Snippet 2:** Python code snippet to generate the panels in \autoref{fig:features}
+![](figures/snippet2.png){width=100%}
 
-```python
-from sarcgraph.sg import SarcGraph
-from sarcgraph.sg_tools import SarcGraphTools
-
-sg = SarcGraph("results", "video")
-sg_tools = SarcGraphTools("results")
-
-# sample_file = "samples/sample_2.avi"
-_, _ = sg.sarcomere_detection("sample_file")
-_ = sg_tools.time_series.sarcomeres_gpr()
-sg_tools._run_all("sample_file")
-
-sg_tools.visualization.spatial_graph()
-sg_tools.visualization.normalized_sarcs_length()
-sg_tools.visualization.F()
-sg_tools.visualization.J()
-```
-
-
-<span style="color: red;"><strong>Snippet 3:</strong></span> Python code snippet to generate the panels in \autoref{fig:validation}
-
-```python
-import numpy as np
-import matplotlib.pyplot as plt
-
-from sarcgraph.sg import SarcGraph
-from sarcgraph.sg_tools import SarcGraphTools
-
-sg = SarcGraph("results", "video")
-sg_tools = SarcGraphTools("results")
-sg_tools_gt = SarcGraphTools("validation-data")
-
-_, _ = sg.sarcomere_detection("samples/sample_1.avi")
-sarcs = sg_tools.time_series.sarcomeres_gpr()
-_ = sg_tools.analysis.compute_F_J()
-lambdas = sg_tools.visualization.F_eigenval_animation(no_anim=True)
-
-sarcs_gt = sg_tools_gt._load_sarcomeres_gpr()
-_ = sg_tools_gt.analysis.compute_F_J()
-lambdas_gt = sg_tools_gt.visualization.F_eigenval_animation(no_anim=True)
-
-sg_tools.visualization.zdiscs_and_sarcs(frame_num=0)
-
-fig, axs = plt.subplots(1, 2, figsize=(10, 5))
-fig.subplots_adjust(wspace=0.3)
-axs[0].grid("on")
-axs[0].set_ylim(-0.1, 0.1)
-axs[0].set_xlabel("frame")
-axs[0].set_ylabel("value")
-for _, data in sarcs.groupby("sarc_id")["length_norm"]:
-    axs[0].plot(
-        data.to_numpy(),
-        linewidth=0.25,
-        color="#942d28",
-        alpha=0.25,
-    )
-axs[0].plot(
-    sarcs.groupby("frame").length_norm.mean(),
-    color="#942d28",
-    label="mean tracked",
-)
-axs[0].plot(
-    sarcs_gt.groupby("frame").length_norm.mean(),
-    "--",
-    color="#999a9d",
-    label="mean ground truth",
-)
-axs[0].legend()
-
-axs[1].grid("on")
-axs[1].set_xlabel("frame")
-axs[1].set_ylabel("value")
-axs[1].plot(
-    lambdas_gt[0, :], color="#a39d9e", label=r"$\lambda_1$ - ground truth"
-)
-axs[1].plot(lambdas[0, :], color="#831915", label=r"$\lambda_1$ - tracked")
-axs[1].plot(
-    lambdas_gt[1, :], ":", color="#182160", label=r"$\lambda_2$ - ground truth"
-)
-axs[1].plot(
-    lambdas[1, :], "--", color="#182160", label=r"$\lambda_2$ - tracked"
-)
-axs[1].legend()
-```
+![](figures/snippet3.png){width=100%}
 
 # References
