### 2023-05-07 Sarc-Graph/sarcgraph#5 @saeedmhz
Thank you for your feedback regarding the testing procedures for our software.

In response to your concerns, we have made changes to the installation process, as outlined in our previous response to Sarc-Graph/sarcgraph#3, which we believe address the issues you raised. Specifically, we have updated the installation instructions to allow users to install the developer version locally from the source code using the editable install command (``pip install -e .``). This ensures that the tests will be run on the source code, rather than on the package installed from the PyPI servers. Therefore, we believe that testing proper installation is not necessary anymore.

We are confident that the updated installation process, coupled with the comprehensive test procedures, will provide a robust testing framework for our software. We hope our revisions adequately address your concerns.

Please let us know if you have any further comments or suggestions.
