### 2023-05-07 5be71c4
adding jupyter notebook youtube guides to the doc
