### 2023-05-09 Sarc-Graph/sarcgraph#6 @saeedmhz
We appreciate your thorough feedback on the manuscript and the opportunity to address the concerns raised. In response to your comments, we have directly implemented the following changes:

1. We introduced a new section in the manuscript titled “SarcGraph vs. Legacy Sarc-Graph,” which discusses the key differences and improvements between the two versions.
2. In the same section, we offer a quantitative comparison of the run time and sarcomere tracking performance for both versions of SarcGraph using an experimental sample video (displayed in Figure 2, left panel) and a synthetically generated sample (depicted in Figure 4).
3. We believe that the addition of specific guides for developer in response to Sarc-Graph/sarcgraph#2 and Sarc-Graph/sarcgraph#3 will better facilitate users interested in extending the software. By customizability, we mostly meant that our code is structured to enable the use of individual steps in isolation or to be replaced with further development. This distinction should be clearer in the revised manuscript.
4. We have incorporated a scale bar in Figure 2, as it was the only plot requiring one.
5. In the appendix, we included code snippets for reproducing the figures 2-4 and added ground truth data for the synthetic sample in Figure 4 to the repository under the ‘validation-data’ folder.
6. We have added a “Statement of Need” section that comprehensively addresses the challenges of the field and the target user group.
7. Lastly, we included a section in the manuscript titled “Current Limitations and Plans for Future Development,” which outlines the software’s limitations and potential future directions.
We hope that these revisions adequately address your concerns, and we welcome any further comments or suggestions. Again, the time taken to provide this comprehensive feedback is much appreciated.

We hope that these revisions adequately address your concerns, and we welcome any further comments or suggestions. Again, the time taken to provide this comprehensive feedback is much appreciated.
