### 2023-05-09 Sarc-Graph/sarcgraph#9 @saeedmhz
Thank you for your valuable and candid feedback regarding the accessibility and user-friendliness of our software for biologists who may be less experienced with programming.

In response to this concern (and some of the other reviewer suggestions), we have made the following changes to improve the user experience and hopefully make our software more accessible:

1. We have simplified the installation process by adding our software to the Conda servers, allowing users to install the software and its dependencies with a single command.
2. To make working with tutorials easier, we have made them accessible through Binder, which allows users to run the Jupyter notebook tutorial files without needing to set up the sarcgraph environment locally.
3. We have added more detailed instructions on how to use the tutorials in the documentation, catering to those who may be less familiar with Jupyter notebooks.
4. Rather than making our own “intro to Jupyter notebook” videos, we identified and now suggest some already existing resources to help new users get started. Specifically, we have provided YouTube links to helpful introductory videos about Jupyter notebooks in our documentation for users who may not be familiar with this tool.

We hope that our revisions adequately address your concerns. Please let us know if you have any further comments or suggestions — it is our sincere intent to make this software accessible to the broader community.
