### 2023-05-09 Sarc-Graph/sarcgraph#8 @saeedmhz
Thank you for your comment regarding the need for a clear "Statement of Need" in our paper. We appreciate the opportunity to address this concern.

We have added a "Statement of Need" section to our manuscript. This new section comprehensively discusses the challenges of the field and the target user group, ensuring that our paper adheres to the JOSS guidelines.
