### 2023-05-09 Sarc-Graph/sarcgraph#7 @saeedmhz
Thank you for your valuable feedback on our paper. We appreciate your suggestions regarding the documentation, and we have made some changes accordingly.

We have addressed the issue you mentioned regarding the installation guide in the repo readme file and the hosted documentation by making sure they match. You can find more information about this in our response to Sarc-Graph/sarcgraph#3.

Additionally, we have removed all the ``$`` signs from the provided commands in the installation guide, making it easier for users to copy and run the commands.
