### 2023-05-09 #5322 @saeedmhz
Dear @Kevin-Mattheus-Moerman, @mbarzegary and @kenatcampbellmusclelab

We would like to express our sincere appreciation for taking the time to review our manuscript and for providing valuable feedback that has helped us improve the quality of our work.

In response to the comments and suggestions provided, we have made revisions to the manuscript and believe that we have addressed all of the issues raised. We hope that the changes we have made have addressed your concerns and improved the clarity and potential impact of this project.

Thank you for your time and effort in reviewing our work — we appreciate the opportunity to submit to JOSS!
