### 2023-05-09 #5322 @saeedmhz
@mbarzegary:

There was one comment that we weren’t totally clear on, specifically “The provided documentation lacks a theory guide. The previous publication and the original software seem to have a more comprehensive guide in this regard. You may consider adding a small appendix to the paper and/or maybe moving part of the previous publication to the documentation as a theory guide.” — to address this comment, in the newly added “statement of need” we make it clear why identifying sarcomeres is a specific challenge, and we point the reader directly to our longer introductory manuscript which is available Open Access (PLOS computational biology) — does this sufficiently address the concern? We would be happy to add more to the manuscript, but we thought it might also be good to avoid repeating ourselves so we were trying to find a balance.
