+++ sarcgraph/sg.py
@@ -228,6 +228,8 @@ class SarcGraph:
         -------
         List[np.ndarray]
 
+        Notes
+        -----
         .. warning::
             Here we use the ``skimage.measure.find_contours`` function to find
             contours. This function returns contours with coordinates in
