+++ setup.py
@@ -5,7 +5,7 @@ with open("README.md", "r") as fh:
 
 setup(
     name="sarcgraph",
-    version="0.1.1",
+    version="0.2.0",
     author="Saeed Mohammadzadeh",
     author_email="saeedmhz@bu.edu",
     description="A software for sarcomere detection and tracking",
