+++ paper.md
@@ -38,17 +38,17 @@ The goal of the SarcGraph software package is to perform sarcomere segmentation,
 
 # SarcGraph in Action
 
-SarcGraph provides users with tools to process images and videos of hiPSC-CMs for z-disc and sarcomere segmentation and tracking. \autoref{fig:sample} demonstrates tracked z-discs and sarcomeres in the first frames of two videos of beating cells. For z-disc and sarcomere segmentation, we build on our previously described work [@zhao2021sarc] and also implement a more efficient sarcomere detection algorithm, detailed in the [Appendix](#appendix). For tracking, we build on a previously developed Python package for particle tracking TrackPy [@allan_daniel_b_2023_7670439]. 
+SarcGraph provides users with tools to process images and movies of hiPSC-CMs for z-disc and sarcomere segmentation and tracking. \autoref{fig:sample} demonstrates tracked z-discs and sarcomeres in the first frames of two movies of beating cells. For z-disc and sarcomere segmentation, we build on our previously described work [@zhao2021sarc] and also implement a more efficient sarcomere detection algorithm, detailed in the [Appendix](#appendix). For tracking, we build on a previously developed Python package for particle tracking TrackPy [@allan_daniel_b_2023_7670439]. 
 
 ![This figure showcases segmented z-discs and detected sarcomeres in the first frame of two beating hiPSC-CM movies. Detected sarcomeres are marked by red stars, while blue contours indicate z-discs. The code to generate this figure is given in Snippet 1 in the [Appendix](#appendix).\label{fig:sample}](figures/sample_vis.png){width=75%}
 
 After initial segmentation and tracking, SarcGraph offers several post-processing analysis and visualization functions. \autoref{fig:features} showcases some of these features. Notably, there are multiple demos and tutorials that further explain these capabilities in the SarcGraph repository.
 
-![This figure presents three panels illustrating some of the post-processing features of the SarcGraph package on a sample video of hiPSC-CMs. Panel (a) displays the spatial graph visualization of the segmented z-discs and sarcomeres. Panel (b) shows the average normalized sarcomere length. Panel (c) depicts the recovered deformation gradient matrix. These results demonstrate the ability of SarcGraph to analyze sarcomere dynamics in hiPSC-CMs by showcasing some of the post-processing features of the SarcGraph package. The code to generate this figure is given in Snippet 2 in the [Appendix](#appendix).\label{fig:features}](figures/features.png){width=95%}
+![This figure presents three panels illustrating some of the post-processing features of the SarcGraph package on a sample movie of hiPSC-CMs. Panel (a) displays the spatial graph visualization of the segmented z-discs and sarcomeres. Panel (b) shows the average normalized sarcomere length. Panel (c) depicts the recovered deformation gradient matrix. These results demonstrate the ability of SarcGraph to analyze sarcomere dynamics in hiPSC-CMs by showcasing some of the post-processing features of the SarcGraph package. The code to generate this figure is given in Snippet 2 in the [Appendix](#appendix).\label{fig:features}](figures/features.png){width=95%}
 
-To validate our methods and ensure correct implementation, we generated challenging synthetic videos with characteristics similar to beating hiPSC-CMs. We used these videos to evaluate the sarcomere detection algorithm by comparing recovered metrics to their known ground truth. \autoref{fig:validation} shows this process for one of the many tested validation examples.
+To validate our methods and ensure correct implementation, we generated challenging synthetic movies with characteristics similar to beating hiPSC-CMs. We used these movies to evaluate the sarcomere detection algorithm by comparing recovered metrics to their known ground truth. \autoref{fig:validation} shows this process for one of the many tested validation examples.
 
[…]
