+++ paper.md
@@ -48,7 +48,7 @@ After initial segmentation and tracking, SarcGraph offers several post-processin
 
 To validate our methods and ensure correct implementation, we generated challenging synthetic movies with characteristics similar to beating hiPSC-CMs. We used these movies to evaluate the sarcomere detection algorithm by comparing recovered metrics to their known ground truth. \autoref{fig:validation} shows this process for one of the many tested validation examples.
 
-![This figure displays the performance of SarcGraph on a synthetically generated movie designed to replicate a beating hiPSC-CM. Panel (a) shows the first frame of the movie with tracked sarcomeres marked by red stars. Panel (b) compares the average normalized sarcomere length obtained from SarcGraph with the ground truth. Panel (c) compares the parameters related to average deformation obtained from SarcGraph with the ground truth. The results demonstrate the validity of the SarcGraph package in accurately analyzing sarcomere dynamics in hiPSC-CMs. As a brief note, the updated version of SarcGraph is better able to recover the ground truth from this synthetic example in comparison to the previous version of the framework showen in Figure 6 in [@zhao2021sarc]. The code to generate this figure is given in Snippet 3 in the [Appendix](#appendix).\label{fig:validation}](figures/validation.png){width=95%}
+![This figure displays the performance of SarcGraph on a synthetically generated movie designed to replicate a beating hiPSC-CM. Panel (a) shows the first frame of the movie with tracked sarcomeres marked by red stars. Panel (b) compares the average normalized sarcomere length obtained from SarcGraph with the ground truth. Panel (c) compares the parameters related to average deformation obtained from SarcGraph with the ground truth. The results demonstrate the validity of the SarcGraph package in accurately analyzing sarcomere dynamics in hiPSC-CMs. As a brief note, the updated version of SarcGraph is better able to recover the ground truth from this synthetic example in comparison to the previous version of the framework shown in Figure 6 in [@zhao2021sarc]. The code to generate this figure is given in Snippet 3 in the [Appendix](#appendix).\label{fig:validation}](figures/validation.png){width=95%}
 
 # SarcGraph vs. Legacy Sarc-Graph
 
@@ -84,7 +84,7 @@ $$ f_3(l_i) = e ^ {-\pi(1 - l_i / l_{avg}) ^ 2} $$
 
 and $c_1$, $c_2$, and $c_3$ are user-defined constants that weigh the importance of each link feature. The default values of $c_k$ and functions for $f_k$ in SarcGraph were selected to produce accurate results on all of the samples that we tested (both real and synthetic data) without any example-specific parameter tuning. However, it is possible to customize these values and functions to suit different scenarios where the defaults may not perform optimally. In future development of SarcGraph, automated parameter tuning will be implemented as needed. 
 
-To prune this spatial graph such that only accurately detected sarcomeres remain, pruning follows four rules: (1) each node can have at most two edges, (2) the angle between edges attached to a node must be greater than a threshold $\theta_{max}$, (3) the edge score must be greater than a threshold $s_{max}$, and (4) When selecting edges to retain, priority is given to those with higher score values. The new algorithm offers improved effectiveness and flexibility over the previous method and is a key component of the new SarcGraph package's sarcomere detection capabilities.
+To prune this spatial graph such that only accurately detected sarcomeres remain, pruning follows four rules: (1) each node can have at most two edges, (2) the angle between edges attached to a node must be greater than a threshold $\theta_{max}$, (3) the edge score must be greater than a threshold $s_{max}$, and (4) when selecting edges to retain, priority is given to those with higher score values. The new algorithm offers improved effectiveness and flexibility over the previous method and is a key component of the new SarcGraph package's sarcomere detection capabilities.
 
[…]
