+++ paper.md
@@ -84,7 +84,7 @@ $$ f_3(l_i) = e ^ {-\pi(1 - l_i / l_{avg}) ^ 2} $$
 
 and $c_1$, $c_2$, and $c_3$ are user-defined constants that weigh the importance of each link feature. The default values of $c_k$ and functions for $f_k$ in SarcGraph were selected to produce accurate results on all of the samples that we tested (both real and synthetic data) without any example-specific parameter tuning. However, it is possible to customize these values and functions to suit different scenarios where the defaults may not perform optimally. In future development of SarcGraph, automated parameter tuning will be implemented as needed. 
 
-To prune this spatial graph to retain the most probable candidates for actual sarcomeres, we adhere to four rules: (1) each node can maintain a maximum of two edges (reflecting the biological reality that a z-disc is typically part of two sarcomeres), (2) all remaining edges must exceed a threshold score, $s_{max}$ (to discard low-probability potential sarcomeres), (3) For nodes with two remaining edges, the angle between them must surpass a threshold, $\theta_{max}$ (which prevents predicted myofibrils from having unrealistic kinked structure), and (4) the retained edges at a node should be those with the highest scores that comply with the first three rules (preserving only the potential sarcomeres that are the most likely to be valid). The new algorithm offers improved effectiveness and flexibility over the previous method and is a key component of the new SarcGraph package's sarcomere detection capabilities.
+To prune this spatial graph to retain the most probable candidates for actual sarcomeres, we adhere to four rules: (1) each node can maintain a maximum of two edges (reflecting the biological reality that a z-disc is typically part of two sarcomeres), (2) all remaining edges must exceed a threshold score, $s_{max}$ (to discard low-probability potential sarcomeres), (3) for nodes with two remaining edges, the angle between them must surpass a threshold, $\theta_{max}$ (which prevents predicted myofibrils from having unrealistic kinked structure), and (4) the retained edges at a node should be those with the highest scores that comply with the first three rules (preserving only the potential sarcomeres that are the most likely to be valid). The new algorithm offers improved effectiveness and flexibility over the previous method and is a key component of the new SarcGraph package's sarcomere detection capabilities.
 
 ![In the initial graph (left) the edges show the potential sarcomeres. After scoring and pruning, the remaining edges that represent the detected sarcomeres are shown in blue lines in the pruned graph (right). The edges that were deleted during pruning are shown in red dashed lines.\label{fig:appendix}](figures/appendix.png){width=95%}
 
