### 2023-05-23 #5322 @saeedmhz
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [x]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.
>
> ## Editor Tasks Prior to Acceptance
> * [ ]  Read the text of the paper and offer comments/corrections (as either a list or a PR)
> * [ ]  Check the references in the paper for corrections (e.g. capitalization)
> * [ ]  Check that the archive title, author list, version tag, and the license are correct
> * [ ]  Set archive DOI with `@editorialbot set <DOI here> as archive`
> * [ ]  Set version with `@editorialbot set <version here> as version`
> * [ ]  Double check rendering of paper with `@editorialbot generate pdf`
> * [ ]  Specifically check the references with `@editorialbot check references` and ask author(s) to update as needed
> * [ ]  Recommend acceptance with `@editorialbot recommend-accept`

@Kevin-Mattheus-Moerman,
In response to these tasks, here are the actions taken:

1. Authors and affiliations, including ORCIDs, have been thoroughly reviewed and confirmed as correct.
2. A new release of the software, v0.2.0, incorporating the latest changes, has been made available on GitHub.
3. This release has been archived on Zenodo with the following DOI: 10.5281/zenodo.7963553.
4. The title and author list in the Zenodo archive perfectly match those in the JOSS paper.
5. The software license for the Zenodo archive aligns with the license as stated in the JOSS paper.

Thank you for your guidance in these matters. Please let me know if there are any other issues that need addressing.
