### 2023-05-23 #5322 @Kevin-Mattheus-Moerman
@saeedmhz @elejeune11 congratulations on this publication!

Thanks a lot for reviewing @kenatcampbellmusclelab and @mbarzegary !
